+++ src/hooks/useRegistry.ts
@@ -14,7 +14,7 @@ const listeners = new Set<() => void>()
 let version = 0
 
 // Notifies the listeners and updates version snapshot
-const notifyRegistryChanged = () => {
+const notifySdkInstanceRegistryChanged = () => {
   version += 1
   listeners.forEach(fn => fn())
 }
@@ -33,37 +33,31 @@ export function useSDKInstanceRegistry(): [SDKRegistry, number] {
   return [sdkInstanceRegistry, version]
 }
 
-export function useConfigRegistry(): [ConfigRegistry, number] {
-  const version = useSyncExternalStore(subscribe, getSnapshot)
-  return [configRegistry, version]
-}
-
 export const getConfigRegistry = () => configRegistry
-
-// Reserve container with a placeholder to prevent other instances from using it
-export const reserveContainer = (key: string) => {
-  sdkInstanceRegistry.set(key, null)
-  notifyRegistryChanged()
-}
 
 export const setConfigInRegistry = (key: string, config: IBeeConfig) => {
   configRegistry.set(key, config)
-  notifyRegistryChanged()
 }
 
 export const removeConfigFromRegistry = (key?: string) => {
   if (key) {
     configRegistry.delete(key)
-    notifyRegistryChanged()
   }
 }
 
 export const setSDKInstanceToRegistry = (key: string, instance: BeefreeSDK | null) => {
   sdkInstanceRegistry.set(key, instance)
-  notifyRegistryChanged()
+  notifySdkInstanceRegistryChanged()
 }
 
 export const removeSDKInstanceFromRegistry = (key: string) => {
   sdkInstanceRegistry.delete(key)
-  notifyRegistryChanged()
+  notifySdkInstanceRegistryChanged()
+}
+
+export const resetRegistry = () => {
+  sdkInstanceRegistry.clear()
+  configRegistry.clear()
+  listeners.clear()
+  version = 0
 }