+++ CHANGELOG.md
@@ -2,6 +2,45 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## 1.1.0-refactor-improvements.0 (2026-01-14)
+
+
+### Features
+
+* add bucket dir prop ([5d76a51](https://github.com/BeefreeSDK/npm-react/commit/5d76a519a46ca02d3905eb12b433c88d5be3b678))
+* App.tsx import EmailBuilder from dist ([f40b97e](https://github.com/BeefreeSDK/npm-react/commit/f40b97e44d90c680068c6f42800b6f99b2c2fc4d))
+* export types for npm package ([0069aaf](https://github.com/BeefreeSDK/npm-react/commit/0069aaf7991b76d6de79f27845596ac409f78a2d))
+* **npm-react:** .env.sample + cleaning up constants ([a921b3b](https://github.com/BeefreeSDK/npm-react/commit/a921b3b26482cf31d3e4d8aed4911a263dbd1228))
+* **npm-react:** adding beefree.io/sdk and fixing some issues with rollup config ([dca5b17](https://github.com/BeefreeSDK/npm-react/commit/dca5b17da2d8568ee21ca3380077087c72a1a1da))
+* **npm-react:** Adding callbacks to props + some changes to the example ([35239e7](https://github.com/BeefreeSDK/npm-react/commit/35239e72527d11caf15ff7a6511b2417d665a9c5))
+* **npm-react:** adding dotenv and eslint conversion to .mjs ([373f114](https://github.com/BeefreeSDK/npm-react/commit/373f114f71d4bfa34b2a24f6e821b3bc033264b0))
+* **npm-react:** adding few libraries (eslint, jest etc.) and adding a sample test ([f506617](https://github.com/BeefreeSDK/npm-react/commit/f50661772db8a4894c8bf3da7ec8d6a4d97e8555))
+* **npm-react:** from context to registry + hook solution ([6b39cad](https://github.com/BeefreeSDK/npm-react/commit/6b39cad0d4925e1eb545f5d4da91f0fcbc352089))
+* **npm-react:** Include BeefreeSDK in the bundle ([674d909](https://github.com/BeefreeSDK/npm-react/commit/674d909afda7bfee42c84b3a45bd9cf291b550fa))
+* **npm-react:** move authentication outside and clean up ([33ff36e](https://github.com/BeefreeSDK/npm-react/commit/33ff36e873a07d6ef17bc3f7d1987699ab2d63fd))
+* **npm-react:** React 19 + shared test ([f5f272f](https://github.com/BeefreeSDK/npm-react/commit/f5f272fdf7e62e1424f293abfbd3c21601c20d1f))
+* **npm-react:** removing ref and introducing back the useRegistry hook ([b641ee0](https://github.com/BeefreeSDK/npm-react/commit/b641ee03fac100cf05260ddba66c5536809a1b80))
+* **npm-react:** rollup config and first components export ([733867d](https://github.com/BeefreeSDK/npm-react/commit/733867d8e60df03babbcee1be24d6ce78d63df15))
+* **npm-react:** start adding example code ([6ba7ade](https://github.com/BeefreeSDK/npm-react/commit/6ba7adeabb97fc087cd5141cb3d0bee24129542d))
+* **npm-react:** trying context as keeper for instance ([fbe761c](https://github.com/BeefreeSDK/npm-react/commit/fbe761c2b81398eb243120c2c9e262d7d6edf2dc))
+* **npm-react:** trying to use ref and loadConfig ([3a592cb](https://github.com/BeefreeSDK/npm-react/commit/3a592cbe88852d8e4024a2791562757a6606c579))
+* **npm-react:** updating beefree.io/sdk dependency ([75ec7a5](https://github.com/BeefreeSDK/npm-react/commit/75ec7a578cfd2786b4a65b1522d4a17b7e89f9e8))
+* **npm-react:** updating example ([3d1a653](https://github.com/BeefreeSDK/npm-react/commit/3d1a6530d173ab8e1dfe072a436a2e703e07425e))
+* **npm-react:** updating the useRegistry hook to verify the registry version before checking builder existence ([544056f](https://github.com/BeefreeSDK/npm-react/commit/544056f3f7a0691e8da7db183da58160be076766))
+* **npm-setup:** initial setup of dependencies. ([1ef4457](https://github.com/BeefreeSDK/npm-react/commit/1ef4457b2ccc52bf6958df1cbc3df8ef27316953))
+* rollup config for example application ([f1c6aee](https://github.com/BeefreeSDK/npm-react/commit/f1c6aee9f5d4d04b3a43be38219fe913952fe0db))
+
+
+### Bug Fixes
+
+* changes the template type ([c2f2d32](https://github.com/BeefreeSDK/npm-react/commit/c2f2d32caf728370a097e74b1008b4d17bdae7b6))
+* environment and bundle ([9312913](https://github.com/BeefreeSDK/npm-react/commit/931291380d26c1ebd6a4f10d5ccd58eb6bf7183f))
+* example type ([b967695](https://github.com/BeefreeSDK/npm-react/commit/b967695a4d68d694ed2ed72fc337e05506b81dd9))
+* export types ([b2eb7ed](https://github.com/BeefreeSDK/npm-react/commit/b2eb7ed60f4463e10476188fa3fe7f3e013a2eeb))
+* revert ([6453eda](https://github.com/BeefreeSDK/npm-react/commit/6453eda489a24a9a64f9d636b531828ddbfebad8))
+* types and url for loader ([4478a14](https://github.com/BeefreeSDK/npm-react/commit/4478a142925505da7841e2b3fc9c5a8055ef22b7))
+* updates app with external types file ([79c500c](https://github.com/BeefreeSDK/npm-react/commit/79c500cb124a240abd4ead3d5842a3c3a78e9007))
+
 ## [1.1.0-refactor-builder-state.13](https://github.com/BeefreeSDK/npm-react/compare/v1.1.0-refactor-builder-state.12...v1.1.0-refactor-builder-state.13) (2025-12-17)